refactor: rename some folders
