+++ src/__tests__/app/[userName]/page.test.tsx
@@ -1,12 +1,12 @@
 import routerMock from 'next-router-mock'
 
 import {makeLink, makeUser, makeUserTheme, setup} from '@/__tests__/__helpers__'
-import * as fetchUser from '@/api/usecases/user'
 import UserPage from '@/app/[username]/page'
-import {Link, User} from '@/models'
+import * as fetchUser from '@/data/usecases/user'
+import {Link, User} from '@/domain/models'
 import {cleanup, screen, waitFor} from '@testing-library/react'
 
-jest.mock('@/api/usecases/user')
+jest.mock('@/data/usecases/user')
 
 jest.mock('next/navigation', () => ({
   ...jest.requireActual('next-router-mock'),
