+++ src/__tests__/app/auth/page.test.tsx
@@ -4,7 +4,7 @@ import {makeFbUser, makeUser, setup} from '@/__tests__/__helpers__'
 import {authStore} from '@/app/auth/context/auth-store'
 import AuthLayout from '@/app/auth/layout'
 import AuthPage from '@/app/auth/page'
-import {User} from '@/models'
+import {User} from '@/domain/models'
 import {parseToUser} from '@/utils/user'
 import {faker} from '@faker-js/faker'
 import {screen, render, cleanup, waitFor} from '@testing-library/react'
