+++ src/app/sitemap.ts
@@ -1,6 +1,6 @@
 import {MetadataRoute} from 'next'
 
-import {fetchAllUsernames} from '@/api/usecases'
+import {fetchAllUsernames} from '@/data/usecases'
 
 export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
   const appUrl = process.env.NEXT_PUBLIC_APP_URL ?? ''
