+++ src/domain/enums/auth.ts
@@ -1,0 +1,5 @@
+export enum Providers {
+  'GITHUB' = 'github.com',
+  'GOOGLE' = 'google.com',
+  'PASSWORD' = 'password',
+}
