+++ src/domain/models/auth.ts
@@ -8,9 +8,3 @@ export type SignUpWithEmailAndPassword = {
   password: string
   displayName: string
 }
-
-export enum Providers {
-  'GITHUB' = 'github.com',
-  'GOOGLE' = 'google.com',
-  'PASSWORD' = 'password',
-}
