+++ src/domain/models/link.ts
