+++ src/domain/models/user.ts
